+++ XamarinApp/LAMA/LAMA/LAMA/Views/CalendarPage.xaml.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using LAMA.ActivityGraphLib;
 
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
